+++ src/process.cpp
@@ -32,5 +32,5 @@ long int Process::UpTime() { return upTime_; }
 // TODO: Overload the "less than" comparison operator for Process objects
 // REMOVE: [[maybe_unused]] once you define the function
 bool Process::operator<(Process const& a) const { 
-    return stof(this->ram_) > stof(a.ram_); 
+    return stoi(this->ram_) > stoi(a.ram_); 
 }
